# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Recruitment Assessment System",
    "description": "AI Recruitment Assessment System with multi-agent evaluation, RAG vector search, and LangGraph workflow.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/jobs": {
      "get": {
        "tags": [
          "1. Job Management"
        ],
        "summary": "List All Jobs",
        "description": "Returns all active job requisitions ordered by most recent.",
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "type": "array",
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "1. Job Management"
        ],
        "summary": "Create Job Requisition",
        "description": "Creates a new job requisition with required and nice-to-have skills.",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/candidates": {
      "get": {
        "tags": [
          "2. Candidate Ingestion & Resume Parsing"
        ],
        "summary": "List Candidates",
        "description": "Lists all candidates. If anonymized is enabled, masks PII for fair evaluation.",
        "operationId": "list_candidates_api_candidates_get",
        "parameters": [
          {
            "name": "anonymized",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Anonymized"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateResponse"
                  },
                  "title": "Response List Candidates Api Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "anonymized_id": {
            "type": "string",
            "title": "Anonymized Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "graduation_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduation Year"
          },
          "education": {
            "items": {},
            "type": "array",
            "title": "Education",
            "default": []
          },
          "experience_years": {
            "type": "integer",
            "title": "Experience Years",
            "default": 0
          },
          "experience": {
            "items": {},
            "type": "array",
            "title": "Experience",
            "default": []
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "projects": {
            "items": {},
            "type": "array",
            "title": "Projects",
            "default": []
          },
          "resume_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Filename"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "anonymized_id",
          "created_at"
        ],
        "title": "CandidateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department",
            "default": "Engineering"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Skills",
            "default": []
          },
          "nice_to_have_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nice To Have Skills",
            "default": []
          },
          "min_experience_years": {
            "type": "integer",
            "title": "Min Experience Years",
            "default": 0
          },
          "min_score_threshold": {
            "type": "number",
            "title": "Min Score Threshold",
            "default": 70.0
          },
          "primary_language": {
            "type": "string",
            "title": "Primary Language",
            "default": "Python"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "title": "JobCreate"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Skills"
          },
          "nice_to_have_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nice To Have Skills"
          },
          "min_experience_years": {
            "type": "integer",
            "title": "Min Experience Years"
          },
          "min_score_threshold": {
            "type": "number",
            "title": "Min Score Threshold"
          },
          "primary_language": {
            "type": "string",
            "title": "Primary Language"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "department",
          "description",
          "required_skills",
          "nice_to_have_skills",
          "min_experience_years",
          "min_score_threshold",
          "primary_language",
          "created_at"
        ],
        "title": "JobResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}